* Fix a number of gcc errors/warnings in KAI.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2404 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/KAI 0.11/DefenseMatrix.cpp
@@ -37,7 +37,7 @@ float3 CDefenseMatrix::GetDefensePos(const UnitDef* def, float3 builderpos)
 {
 	ai->ut->UpdateChokePointArray();
 	int f3multiplier = 8*THREATRES;
-	int Range = ai->ut->GetMaxRange(def) / f3multiplier;
+	int Range = int(ai->ut->GetMaxRange(def) / f3multiplier);
 	int squarerange = Range * Range;
 	float bestscore = 0;
 	int bestspotx = 0;
@@ -232,7 +232,7 @@ void CDefenseMatrix::AddDefense(float3 pos, const UnitDef* def)
 {
 	
 	int f3multiplier = 8*THREATRES;
-	int Range = ai->ut->GetMaxRange(def) / f3multiplier;
+	int Range = int(ai->ut->GetMaxRange(def) / f3multiplier);
 	int squarerange = Range * Range;
 	int x,y;
 	ai->math->F32XY(pos,&x,&y,8);
@@ -242,7 +242,7 @@ void CDefenseMatrix::AddDefense(float3 pos, const UnitDef* def)
 	for (int myx = x - Range; myx <= x + Range; myx++){
 		if (myx >= 0 && myx < ai->pather->PathMapXSize){
 			for (int myy = y - Range; myy <= y + Range; myy++){
-				int distance = (x - myx)*(x - myx) + (y - myy)*(y - myy) - 0.5;
+				int distance = int((x - myx)*(x - myx) + (y - myy)*(y - myy) - 0.5);
 				if (myy >= 0 && myy < ai->pather->PathMapYSize && (distance) <= squarerange){
 					for(int i = 0; i < ai->pather->NumOfMoveTypes;i++){
 						ChokeMapsByMovetype[i][myy * ai->pather->PathMapXSize + myx] /= 2;
@@ -261,7 +261,7 @@ void CDefenseMatrix::AddDefense(float3 pos, const UnitDef* def)
 void CDefenseMatrix::RemoveDefense(float3 pos, const UnitDef* def)
 {
 	int f3multiplier = 8*THREATRES;
-	int Range = ai->ut->GetMaxRange(def) / f3multiplier;
+	int Range = int(ai->ut->GetMaxRange(def) / f3multiplier);
 	int squarerange = Range * Range;
 	int x,y;
 	ai->math->F32XY(pos,&x,&y,8);
@@ -270,7 +270,7 @@ void CDefenseMatrix::RemoveDefense(float3 pos, const UnitDef* def)
 	for (int myx = x - Range; myx <= x + Range; myx++){
 		if (myx >= 0 && myx < ai->pather->PathMapXSize){
 			for (int myy = y - Range; myy <= y + Range; myy++){
-				int distance = (x - myx)*(x - myx) + (y - myy)*(y - myy) - 0.5;
+				int distance = int((x - myx)*(x - myx) + (y - myy)*(y - myy) - 0.5);
 				if (myy >= 0 && myy < ai->pather->PathMapYSize && (distance) <= squarerange){
 					for(int i = 0; i < ai->pather->NumOfMoveTypes;i++){
 						ChokeMapsByMovetype[i][myy * ai->pather->PathMapXSize + myx] *= 2;
