+++ AI/Global/KAI 0.11/MetalMap.cpp
@@ -15,7 +15,7 @@ CMetalMap::CMetalMap(AIClasses* ai)
 	MetalMapHeight =ai->cb->GetMapHeight() / 2; //metal map has 1/2 resolution of normal map
 	MetalMapWidth =ai->cb->GetMapWidth() / 2;
 	TotalCells = MetalMapHeight * MetalMapWidth;
-	XtractorRadius =ai->cb->GetExtractorRadius()/ 16;
+	XtractorRadius = int(ai->cb->GetExtractorRadius() / 16);
 	DoubleRadius = XtractorRadius * 2;
 	SquareRadius = XtractorRadius * XtractorRadius; //used to speed up loops so no recalculation needed
 	DoubleSquareRadius = DoubleRadius * DoubleRadius; // same as above 
@@ -95,7 +95,7 @@ void CMetalMap::GetMetalPoints()
 	for (int a=0;a<DoubleRadius+1;a++){ 
 		float z=a-XtractorRadius;
 		float floatsqrradius = SquareRadius;
-		xend[a]=sqrt(floatsqrradius-z*z);
+		xend[a]=int(sqrtf(floatsqrradius-z*z));
 	}
 	//Load up the metal Values in each pixel
 	const unsigned char *metalMapArray = ai->cb->GetMetalMap();
